# docs/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GW2 WvW日志系统",
    "description": "GW2 WvW日志解析与评分系统后端API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ei-analysis/{log_id}": {
      "get": {
        "tags": [
          "EI分析"
        ],
        "summary": "获取日志战斗摘要",
        "description": "获取日志的战斗摘要数据。\n返回精简版数据（~9KB），替代原先 40MB 完整 EI JSON。\nv2.1 增加聚合统计、Buff 排行、支援数据。",
        "operationId": "get_ei_summary_api_v1_ei_analysis__log_id__get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Log Id"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "排序字段",
              "default": "damage",
              "title": "Sort By"
            },
            "description": "排序字段"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ei-analysis/{log_id}/player/{account}/rotation": {
      "get": {
        "tags": [
          "EI分析"
        ],
        "summary": "获取玩家技能循环",
        "description": "获取指定玩家在指定日志中的技能循环(rotation)和技能释放次数。\n返回 EiPlayer.rotation_json + stats_all_json 中的 skill_casts + EiSkillMap 技能映射。",
        "operationId": "get_player_rotation_api_v1_ei_analysis__log_id__player__account__rotation_get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Log Id"
            }
          },
          {
            "name": "account",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ei-analysis/{log_id}/player/{account}": {
      "get": {
        "tags": [
          "EI分析"
        ],
        "summary": "获取玩家在某日志中的详细数据",
        "description": "获取指定玩家在指定日志中的详细统计数据。",
        "operationId": "get_player_detail_api_v1_ei_analysis__log_id__player__account__get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Log Id"
            }
          },
          {
            "name": "account",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "操作是否成功"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示消息"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code",
            "description": "错误代码"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "时间戳"
          },
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code",
            "description": "HTTP状态码"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ApiResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}